+++ man-roxygen/ex-map_continent.R
@@ -11,16 +11,12 @@ if (interactive()) {
   )
   server <- function(input, output) {
 
-    orig_data <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))$orig_data
+    DATA <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))
+    orig_data_aggregate <- DATA$orig_data_aggregate
 
-    pop_data = get_pop_datahub()
-    orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+    countries_data_map = DATA$countries_data_map
 
     #countries_data_map <- Covid19Mirai:::load_countries_datahub_map(destpath = system.file("./countries_data", package = "Covid19Mirai"))
-    rds_map = "WorldMap_sp_spl.rds"
-    message("read map from RDS ", rds_map)
-    countries_data_map = readRDS(file =  file.path(system.file("./countries_data", package = "Covid19Mirai"),rds_map))
-
 
     orig_data_aggregate_cont <-
       orig_data_aggregate %>% filter(continent == cont)