+++ man-roxygen/ex-mod_compare_nth_cases_plot.R
@@ -10,10 +10,10 @@
     )
     server <- function(input, output, session) {
       # Data ----
-      orig_data <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))$orig_data
+      orig_data_with_ch = readRDS(file =  file.path(system.file("./datahub", package = "Covid19Mirai"),rds_data))
 
-      pop_data = get_pop_datahub()
-      orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+      orig_data_aggregate       = orig_data_with_ch$orig_data_aggregate
+      orig_data_ch_2  = orig_data_with_ch$orig_data_ch_2
 
       nn = 1000;
      #  data_filtered <- #reactive({
@@ -36,11 +36,8 @@ if (interactive()) {
   )
   server <- function(input, output, session) {
     # Data ----
-    orig_data <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))$orig_data
-
-
-    pop_data = get_pop_datahub()
-    orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+    DATA <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))
+    orig_data_aggregate <- DATA$orig_data_aggregate
 
     n = 1000;
 
@@ -67,11 +64,8 @@ if (interactive()) {
   )
   server <- function(input, output, session) {
     # Data ----
-    orig_data <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))$orig_data
-
-
-    pop_data = get_pop_datahub()
-    orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+    DATA <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))
+    orig_data_aggregate <- DATA$orig_data_aggregate
 
     n = 1000
     countries = c("Switzerland")
@@ -136,11 +130,8 @@ if (interactive()) {
   )
   server <- function(input, output, session) {
     # Data ----
-    orig_data <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))$orig_data
-
-
-    pop_data = get_pop_datahub()
-    orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+    DATA <- readRDS(system.file("datahub/DATA.rds", package = "Covid19Mirai"))
+    orig_data_aggregate <- DATA$orig_data_aggregate
 
     n = 1000
     countries_data <-